+++ app/simulations/CMakeLists.txt
@@ -1,7 +1,9 @@
-declare_program(name IsoSteadyState
+declare_program(name IsoSteadyStateSim
                 dependencies Boost::serialization
                 MaierSaupe
-                BoundaryValues)
+                BoundaryValues
+                Postprocessors
+                LiquidCrystalSystems)
 
 # not compiling, IDK why
 #declare_program(name LdGIsoSteadyState
