+++ src/CMakeLists.txt
@@ -1,4 +1,4 @@
-find_package(Boost COMPONENTS serialization REQUIRED)
+find_package(Boost COMPONENTS serialization program_options REQUIRED)
 
 # helper library for Sphere Lebedev Rules
 add_subdirectory(sphere_lebedev_rule)
@@ -11,6 +11,9 @@ add_subdirectory(BoundaryValues)
 
 # postprocessors library
 add_subdirectory(Postprocessors)
+
+# liquid crystal systems library
+add_subdirectory(LiquidCrystalSystems)
 
 # TODO: break up maier-saupe library into other, smaller libraries
 # shared library for Maier Saupe Liquid Crystal simulations
