+++ src/LiquidCrystalSystems/IsoSteadyState.inst
@@ -1,0 +1,70 @@
+#ifndef ISO_STEADY_STATE_INST
+#define ISO_STEADY_STATE_INST
+
+template class IsoSteadyState<2, 6>;
+template class IsoSteadyState<2, 14>;
+template class IsoSteadyState<2, 26>;
+template class IsoSteadyState<2, 38>;
+template class IsoSteadyState<2, 50>;
+template class IsoSteadyState<2, 74>;
+template class IsoSteadyState<2, 86>;
+template class IsoSteadyState<2, 110>;
+template class IsoSteadyState<2, 146>;
+template class IsoSteadyState<2, 170>;
+template class IsoSteadyState<2, 194>;
+template class IsoSteadyState<2, 230>;
+template class IsoSteadyState<2, 266>;
+template class IsoSteadyState<2, 302>;
+template class IsoSteadyState<2, 350>;
+template class IsoSteadyState<2, 434>;
+template class IsoSteadyState<2, 590>;
+template class IsoSteadyState<2, 770>;
+template class IsoSteadyState<2, 974>;
+template class IsoSteadyState<2, 1202>;
+template class IsoSteadyState<2, 1454>;
+template class IsoSteadyState<2, 1730>;
+template class IsoSteadyState<2, 2030>;
+template class IsoSteadyState<2, 2354>;
+template class IsoSteadyState<2, 2702>;
+template class IsoSteadyState<2, 3074>;
+template class IsoSteadyState<2, 3470>;
+template class IsoSteadyState<2, 3890>;
+template class IsoSteadyState<2, 4334>;
+template class IsoSteadyState<2, 4802>;
+template class IsoSteadyState<2, 5294>;
+template class IsoSteadyState<2, 5810>;
+
+template class IsoSteadyState<3, 6>;
+template class IsoSteadyState<3, 14>;
+template class IsoSteadyState<3, 26>;
+template class IsoSteadyState<3, 38>;
+template class IsoSteadyState<3, 50>;
+template class IsoSteadyState<3, 74>;
+template class IsoSteadyState<3, 86>;
+template class IsoSteadyState<3, 110>;
+template class IsoSteadyState<3, 146>;
+template class IsoSteadyState<3, 170>;
+template class IsoSteadyState<3, 194>;
+template class IsoSteadyState<3, 230>;
+template class IsoSteadyState<3, 266>;
+template class IsoSteadyState<3, 302>;
+template class IsoSteadyState<3, 350>;
+template class IsoSteadyState<3, 434>;
+template class IsoSteadyState<3, 590>;
+template class IsoSteadyState<3, 770>;
+template class IsoSteadyState<3, 974>;
+template class IsoSteadyState<3, 1202>;
+template class IsoSteadyState<3, 1454>;
+template class IsoSteadyState<3, 1730>;
+template class IsoSteadyState<3, 2030>;
+template class IsoSteadyState<3, 2354>;
+template class IsoSteadyState<3, 2702>;
+template class IsoSteadyState<3, 3074>;
+template class IsoSteadyState<3, 3470>;
+template class IsoSteadyState<3, 3890>;
+template class IsoSteadyState<3, 4334>;
+template class IsoSteadyState<3, 4802>;
+template class IsoSteadyState<3, 5294>;
+template class IsoSteadyState<3, 5810>;
+
+#endif
